удаляет более не нужные файлы
